Tests/LibWeb/Text/expected/UIEvents/label-double-click-toggles-twice.txt

initial: false
after single click: true
after double click: true
